resume_builder/requirements.txt
Django==4.2.21
djangorestframework==3.15.2
django-cors-headers==4.4.0
whitenoise==6.7.0
python-dotenv==1.0.1
reportlab==4.2.2
pikepdf==10.5.0
twilio==9.2.3
psycopg[binary]==3.2.13
Pillow==10.4.0
gunicorn==23.0.0
